requirements.txt
Django==2.1.5
psycopg2-binary==2.7.7
django-confy==1.0.4
django-bootstrap3==10.0.1
openpyxl==2.4.2
django-reversion==3.0.2
pytz==2018.9
django-timezone-field==2.0
requests==2.20.0
Unipath==1.1
six==1.10
python-dateutil==2.6.0
future==0.16.0
unicodecsv==0.14.1

# frictionless datapackage and tableschema
datapackage==1.2.6
tableschema==1.0.13

# rest API
djangorestframework==3.8.2
coreapi==2.3.3
# next line is to fix a strange dep error from (the openapi-codec==1.3.1 create issues)
django-filter==2.1.0
django-crispy-forms==1.6.1
django-cors-headers==1.3.1
dry-rest-permissions==0.1.8
djangorestframework-gis==0.14
drf-yasg==1.7.4
django-extra-fields==1.0.0
Pillow==5.1.0
djoser==1.2.1

# for S3 static/media storage
django-storages==1.7.1
boto3==1.9.91

# dbca for SSO login
git+https://github.com/dbca-wa/dbca-utils@1.0.0

# TODO: to get rid of
django-grappelli==2.8.2


# suspicious dependencies
dj-static==0.0.6

# test, deploy and dev
Fabric>=1.10.1,<2.0
coverage>=4.2,<5.0
flake8>=2.4.0,<3.0
mixer==6.1.3
django-extensions==2.1.5
ipython==7.2.0
Werkzeug==0.14.1
factory-boy==2.11.1
gunicorn==19.9.0